+++ src/main.js
@@ -23,6 +23,7 @@ export { default as DiscloseTable } from './components/DiscloseTable';
 export { default as Dropdown } from './components/Dropdown';
 export { default as EditableInput } from './components/EditableInput';
 export { default as EmptyDashboard } from './components/EmptyDashboard';
+export { default as Fieldset } from './components/Fieldset';
 export { default as FilterPicker } from './components/FilterPicker';
 export { default as HelpPopover } from './components/HelpPopover';
 export { default as ButtonIcon } from './components/ButtonIcon';