+++ CHANGELOG.md
@@ -5,11 +5,12 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This file is similar to the format suggested by [Keep a CHANGELOG](https://github.com/olivierlacan/keep-a-changelog).
 
 ## Unreleased
+- [Chore] Export **Fieldset** component for now, full docs and stories to come. ([#1198](https://github.com/optimizely/oui/pull/1198))
 
 - [Patch] Fix **Card** story, relax prop-type check for children ([#1194](https://github.com/optimizely/oui/pull/1194))
 
 ## 44.2.1 - 2019-07-16
-- [Chore] Upgrade react-oui-icons to v2.6.1. (#1193)
+- [Chore] Upgrade react-oui-icons to v2.6.1. ([#1193](https://github.com/optimizely/oui/pull/1193))
 
 ## 44.2.0 - 2019-07-12
 - [Feature] Add **CodeDiff** component ([#1191](https://github.com/optimizely/oui/pull/1191))